// src/types/dmm-keywordpage.ts
export const DMMItemSchema = z.object({
	service_code: z.string(),
	service_name: z.string(),
	floor_code: z.string(),
	floor_name: z.string(),
	category_name: z.string(),
	content_id: z.string(),
	product_id: z.string(),
	title: z.string(),
	volume: z.string().optional(),
	number: z.string().optional(),
	review: z
		.object({
			count: z.coerce.number(), // 数値型として定義
			average: z.coerce.number(), // 数値型として定義
		})
		.optional()
		.nullable(),
	URL: z.string(),
	affiliateURL: z.string(),
	imageURL: ImageURLSchema.optional().nullable(),
	tachiyomi: z
		.object({
			URL: z.string(),
			affiliateURL: z.string(),
		})
		.optional()
		.nullable(),
	sampleImageURL: SampleImageURLSchema.optional().nullable(),
	sampleMovieURL: SampleMovieURLSchema.optional().nullable(),
	prices: z
		.object({
			price: z.string(),
			list_price: z.string(),
			deliveries: z.object({
				delivery: z.array(
					z.object({
						type: z.string(),
						price: z.string(),
						list_price: z.string(),
					}),
				),
			}),
		})
		.optional()
		.nullable(),
	date: z.string().optional().nullable(),
	iteminfo: ItemInfoSchema.optional().nullable(),
	cdinfo: CDInfoSchema.optional().nullable(),
	jancode: z.string().optional().nullable(),
	maker_product: z.string().optional().nullable(),
	isbn: z.string().optional().nullable(),
	stock: z.string().optional().nullable(),
	directory: z
		.array(
			z.object({
				id: z.coerce.number(), // 数値型として定義
				name: z.string(),
			}),
		)
		.optional()
		.nullable(),
	campaign: z.array(CampaignSchema).optional().nullable(),
	db_id: z.coerce.number(), // 数値型として定義
})

// APIレスポンス全体のスキーマ
export const GetKVTop100ResponseSchema = z.object({
	keyword: z.string(),
	items: z.array(DMMItemSchema),
	createdAt: z.string(),
})
// ...
export type GetKVTop100Response = z.infer<typeof GetKVTop100ResponseSchema>